Use ButtonText base class for buttons with text
